rename module from 'k-openssl-common' to 'crypto-common'
