+++ kotlin/concepts/authentication/concept-encryption-key/build.gradle
@@ -6,5 +6,5 @@ plugins {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    api project(path: ':kotlin:crypto:k-openssl-common')
+    api project(path: ':kotlin:crypto:crypto-common')
 }
