+++ kotlin/concepts/authentication/concept-encryption-key/src/main/java/io/matthewnelson/concept_encryption_key/EncryptionKey.kt
@@ -15,7 +15,7 @@
 * */
 package io.matthewnelson.concept_encryption_key
 
-import io.matthewnelson.k_openssl_common.clazzes.Password
+import io.matthewnelson.crypto_common.clazzes.Password
 
 class EncryptionKey private constructor(val privateKey: Password, val publicKey: Password) {
     companion object {
