+++ kotlin/crypto/crypto-common/src/main/java/io/matthewnelson/crypto_common/annotations/UnencryptedDataAccess.kt
@@ -13,8 +13,8 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.k_openssl_common.annotations
+package io.matthewnelson.crypto_common.annotations
 
 @RequiresOptIn(level = RequiresOptIn.Level.ERROR)
 @Retention(AnnotationRetention.BINARY)
-annotation class RawPasswordAccess
+annotation class UnencryptedDataAccess
