+++ kotlin/crypto/crypto-common/src/main/java/io/matthewnelson/crypto_common/clazzes/EncryptedString.kt
@@ -13,10 +13,6 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.k_openssl_common.exceptions
+package io.matthewnelson.crypto_common.clazzes
 
-class EncryptionException: Exception {
-    constructor(cause: Throwable?) : super(cause) {}
-    constructor(message: String?) : super(message) {}
-    constructor(message: String?, cause: Throwable) : super(message, cause) {}
-}
+inline class EncryptedString(val value: String)
