+++ kotlin/crypto/crypto-common/src/main/java/io/matthewnelson/crypto_common/clazzes/Password.kt
@@ -13,9 +13,9 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.k_openssl_common.clazzes
+package io.matthewnelson.crypto_common.clazzes
 
-import io.matthewnelson.k_openssl_common.annotations.RawPasswordAccess
+import io.matthewnelson.crypto_common.annotations.RawPasswordAccess
 
 @Suppress("NOTHING_TO_INLINE")
 @OptIn(RawPasswordAccess::class)
