+++ kotlin/crypto/crypto-common/src/main/java/io/matthewnelson/crypto_common/extensions/Extensions.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.k_openssl_common.extensions
+package io.matthewnelson.crypto_common.extensions
 
 import java.nio.ByteBuffer
 import java.nio.CharBuffer
