+++ kotlin/crypto/k-openssl/build.gradle
@@ -6,7 +6,7 @@ plugins {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(path: ':kotlin:crypto:bouncycastle-ktx')
-    api project(path: ':kotlin:crypto:k-openssl-common')
+    api project(path: ':kotlin:crypto:crypto-common')
     implementation project(path: ':kotlin:encoders:base64')
     implementation KAdeps.kotlin.coroutinesCore
