+++ kotlin/crypto/k-openssl/src/test/java/io/matthewnelson/k_openssl/KOpenSSLUnitTest.kt
@@ -15,12 +15,12 @@
 * */
 package io.matthewnelson.k_openssl
 
-import io.matthewnelson.k_openssl_common.annotations.UnencryptedDataAccess
+import io.matthewnelson.crypto_common.annotations.UnencryptedDataAccess
 import io.matthewnelson.k_openssl.algos.AES256CBC_PBKDF2_HMAC_SHA256
-import io.matthewnelson.k_openssl_common.clazzes.EncryptedString
-import io.matthewnelson.k_openssl_common.clazzes.HashIterations
-import io.matthewnelson.k_openssl_common.clazzes.Password
-import io.matthewnelson.k_openssl_common.clazzes.UnencryptedString
+import io.matthewnelson.crypto_common.clazzes.EncryptedString
+import io.matthewnelson.crypto_common.clazzes.HashIterations
+import io.matthewnelson.crypto_common.clazzes.Password
+import io.matthewnelson.crypto_common.clazzes.UnencryptedString
 import io.matthewnelson.test_k_openssl.OpenSSLTestHelper
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.resetMain
