+++ kotlin/test/test-feature-authentication-core/src/main/java/io/matthewnelson/test_feature_authentication_core/TestEncryptionKeyHandler.kt
@@ -18,9 +18,9 @@ package io.matthewnelson.test_feature_authentication_core
 import io.matthewnelson.concept_encryption_key.EncryptionKey
 import io.matthewnelson.concept_encryption_key.EncryptionKeyException
 import io.matthewnelson.concept_encryption_key.EncryptionKeyHandler
-import io.matthewnelson.k_openssl_common.annotations.RawPasswordAccess
-import io.matthewnelson.k_openssl_common.clazzes.HashIterations
-import io.matthewnelson.k_openssl_common.clazzes.Password
+import io.matthewnelson.crypto_common.annotations.RawPasswordAccess
+import io.matthewnelson.crypto_common.clazzes.HashIterations
+import io.matthewnelson.crypto_common.clazzes.Password
 
 /**
  * Extend and implement your own overrides if desired.
