+++ settings.gradle
@@ -35,8 +35,8 @@ include ':kotlin:concepts:concept-views'
 /// :kotlin:crypto ///
 //////////////////////
 include ':kotlin:crypto:bouncycastle-ktx'
+include ':kotlin:crypto:crypto-common'
 include ':kotlin:crypto:k-openssl'
-include ':kotlin:crypto:k-openssl-common'
 
 ////////////////////////
 /// :kotlin:encoders ///
